User find by id method created

+++ src/main/java/com/odev/yemektarifiodevi/controller/UserController.java
@@ -34,7 +34,7 @@ public class UserController {
 
 
     @GetMapping("/{id}")
-    public User getUsers(@PathVariable int id) {
+    public User getUsers(@PathVariable Long id) {
         return userService.getByID(id);
